+++ lib/rubocop/cop/rails/action_filter.rb
@@ -12,11 +12,37 @@ module RuboCop
 
         MSG = 'Prefer `%s` over `%s`.'
 
-        FILTER_METHODS = [:before_filter, :skip_before_filter,
-                          :after_filter, :around_filter]
+        FILTER_METHODS = [
+          :after_filter,
+          :append_after_filter,
+          :append_around_filter,
+          :append_before_filter,
+          :around_filter,
+          :before_filter,
+          :prepend_after_filter,
+          :prepend_around_filter,
+          :prepend_before_filter,
+          :skip_after_filter,
+          :skip_around_filter,
+          :skip_before_filter,
+          :skip_filter
+        ]
 
-        ACTION_METHODS = [:before_action, :skip_before_action,
-                          :after_action, :around_action]
+        ACTION_METHODS = [
+          :after_action,
+          :append_after_action,
+          :append_around_action,
+          :append_before_action,
+          :around_action,
+          :before_action,
+          :prepend_after_action,
+          :prepend_around_action,
+          :prepend_before_action,
+          :skip_after_action,
+          :skip_around_action,
+          :skip_before_action,
+          :skip_action_callback
+        ]
 
         def on_block(node)
           method, _args, _body = *node