+++ spec/rubocop/cop/rails/action_filter_spec.rb
@@ -6,6 +6,46 @@ describe RuboCop::Cop::Rails::ActionFilter, :config do
   subject(:cop) { described_class.new(config) }
   let(:cop_config) { { 'Include' => nil } }
 
+  describe '::FILTER_METHODS' do
+    it 'contains all of the filter methods' do
+      expect(described_class::FILTER_METHODS).to eq([
+        :after_filter,
+        :append_after_filter,
+        :append_around_filter,
+        :append_before_filter,
+        :around_filter,
+        :before_filter,
+        :prepend_after_filter,
+        :prepend_around_filter,
+        :prepend_before_filter,
+        :skip_after_filter,
+        :skip_around_filter,
+        :skip_before_filter,
+        :skip_filter
+      ])
+    end
+  end
+
+  describe '::ACTION_METHODS' do
+    it 'contains all of the action methods' do
+      expect(described_class::ACTION_METHODS).to eq([
+        :after_action,
+        :append_after_action,
+        :append_around_action,
+        :append_before_action,
+        :around_action,
+        :before_action,
+        :prepend_after_action,
+        :prepend_around_action,
+        :prepend_before_action,
+        :skip_after_action,
+        :skip_around_action,
+        :skip_before_action,
+        :skip_action_callback
+      ])
+    end
+  end
+
   context 'when style is action' do
     before do
       cop_config.update('EnforcedStyle' => 'action')